Add files via upload
This update has functions to print the Binary Search Tree using Post Order Traversal. This was very challenging to implement from scratch using iteration. Each side of the root is treated separately. First the height of each sub tree "side" is found by traversing down left always then right always until a null left or right pointer is found. Then starting from the sub root we traverse down until we are one level away from the null row. then we try to print the left and right pointers of our current node, then we subtract 1 from the height. We keep doing this until the "height" has been reduced to zero, effectively advancing "up" the tree each time. Then we print the corresponding sub root and then finally the root.

Thoughts:
While Testing this it would have been advantageous to only need to enter the key to make a new entry. and then fill in the details of each entry (name, address, ect.) after the infrastructure was built. This would have saved a lot of time. Also It became apparent that searching for an entry will be difficult given the randomized aspect of the key. It may end up that to search, first the tree is printed and then the user is prompted to enter the key of their desired entry. Timeline allowing, a map may be used to let a user input a name and return the key of that name to be seached by. Also to meet the timeline, In Order traversal may be done using recursion.